Add files via upload
Simple cylindrical shape was used for easier connection between the forearm and the arm itself. The arm for poppy robot open source files was used and to consider the holes it has at the end, 4 holes were made in the arm-connector part. 

One thing is missing in my design and can be noticed is the shape or the way these pins (screws place) were organized should be changed to be able to assemble the arm.