Purpose of this:
Depending on the selected motor, the design of this part on the side can be changed. For now, as for the first task, a part from the open source files for Poppy robot was taken and cut into half. The purpose of using it is to connect the forearm with the upper parts and to hold the motor in between to provide flexible movement.