Design explanation:
The joint type used here is the one opposite to centre rational joint, in which it designed to cover the shaft of the motor for easier movement. One important factor was kept in mind while designing this was the distance from the joint to the motor. The shorter the distance is, the better and hight torque.