Idea- source:
The idea of the elbow joint can be seen as in the uploaded pictures which were taken from one of the online course. In my design, the joint was attached to the forearm using a small screw and the motor would be placed vertically so that its shaft would control the arm's movement.