Add files via upload
The elbow joint was make this way to simplify the robot movement. The holes on the sides were kept to provide the selection option for the user to choose where to place the motor's shaft. 
As the purpose of this tasks, as mentioned before, was to familiarize myself with the program and main points, dimensions were not 100% accurate for now.