Design Explanation:
As mentioned before, to keep the robotic vide, blocks were used in the design. The block for upper arm was emptied from inside to a specific level. This was to consider the location of the motor that should be placed to the shoulder part. Holes from upper and lower sides were make to keep the servos in place and prevent them from unwanted movement. 
From top and side views, names can be seen which used for decoration purpose.